Move language macro
