+++ src/macros.rs
@@ -1,4 +1,44 @@
 #[macro_export]
+macro_rules! language_enum {
+    ([$($language:ident),*]) => {
+        #[derive(Debug, Serialize, Deserialize, Copy, Clone, PartialEq, Hash, Eq)]
+        #[allow(non_camel_case_types)]
+        pub enum Language {
+            $( $language, )*
+        }
+
+        impl Language {
+            pub fn all() -> &'static [Language] {
+                static ALL: &[Language] = &[$( Language::$language ),*];
+                ALL
+            }
+        }
+
+        impl ::std::str::FromStr for Language {
+            type Err=::failure::Error;
+            fn from_str(s: &str) -> ::std::result::Result<Language, Self::Err> {
+                match &*s.to_uppercase() {
+                    $(
+                        stringify!($language) => Ok(Language::$language),
+                    )*
+                    _ => bail!("Unknown language: {}", s)
+                }
+            }
+        }
+
+        impl ::std::string::ToString for Language {
+            fn to_string(&self) -> String {
+                match self {
+                    $(
+                        &Language::$language => stringify!($language).to_lowercase(),
+                    )*
+                }
+            }
+        }
+    }
+}
+
+#[macro_export]
 macro_rules! enum_kind {
     ($kindname:ident, [$($varname:ident),*]) => {
         #[derive(Debug, Serialize, Deserialize, Copy, Clone, PartialEq, Hash, Eq)]
